refactor(magazijn,common): NL-standaarden review fixes
- max-body-size 2M → 40M zodat een aanlever-request met 25 MiB bijlage
  (~33 MiB base64) past; eerdere limit gaf 413 vóór validatie.
- V3 partial index `idx_berichten_ontvanger_actief` voor de lijst-query;
  zonder index seq-scan + sort op groeiende tabel.
- BijlageRepository.toBijlage dereferentieert niet langer de LAZY
  bericht-associatie; krijgt de bekende `berichtId` als parameter →
  geen LazyInitializationException als de Bijlage gedetacheerd raakt.
- `autoriseerOntvanger` duplicatie tussen BerichtOphaalService en
  BerichtBeheerService gefactord naar `opslag/BerichtAutorisatie`.
- Quarkus globale security-headers in beide services zodat ook
  /openapi.json, /q/health en /q/metrics ze krijgen (de JAX-RS
  SecurityHeadersFilter dekt alleen JAX-RS-paden).
- /openapi.json expliciet geconfigureerd op berichtenmagazijn
  (ADR-vereiste — was alleen sessiecache).
- LdvEndpointValidator verhuisd naar fbs-common zodat ook
  berichtensessiecache TLS-afdwinging in %prod krijgt (BIO 13.2.1).
