+++ services/berichtenmagazijn/src/main/resources/application.properties
@@ -3,9 +3,24 @@ quarkus.application.version=0.1.0
 
 quarkus.http.port=8090
 
-# Max request body: ruim boven Bericht.MAX_INHOUD_BYTES (1 MiB) zodat
-# JSON-envelope en escaping passen. Grotere requests worden door Quarkus afgewezen.
-quarkus.http.limits.max-body-size=2M
+# OpenAPI / Swagger UI (ADR: spec MOET op /openapi.json beschikbaar zijn)
+quarkus.smallrye-openapi.path=/openapi.json
+quarkus.swagger-ui.always-include=true
+
+# Security headers op ÁLLE HTTP-responses (ook /openapi.json, /q/health,
+# /q/metrics, dev-UI) — niet alleen op JAX-RS-paden. De SecurityHeadersFilter
+# in fbs-common dekt JAX-RS responses; deze Quarkus-config dekt de rest.
+quarkus.http.header."X-Frame-Options".value=DENY
+quarkus.http.header."X-Content-Type-Options".value=nosniff
+quarkus.http.header."Strict-Transport-Security".value=max-age=31536000; includeSubDomains; preload
+quarkus.http.header."Content-Security-Policy".value=frame-ancestors 'none'
+quarkus.http.header."Referrer-Policy".value=no-referrer
+
+# Max request body: ruim boven Bericht.MAX_INHOUD_BYTES (1 MiB) + één
+# bijlage van Bijlage.MAX_CONTENT_BYTES (25 MiB) + base64-overhead (~33%)
+# + JSON-envelope. 40M dekt een aanlever-request met inhoud en bijlage(n);
+# grotere requests worden door Quarkus afgewezen.
+quarkus.http.limits.max-body-size=40M
 
 # Persistentie — PostgreSQL voor alle profielen, zodat dev/test/prod hetzelfde
 # DBMS-gedrag laten zien. Dev gebruikt de Postgres-service uit `compose.yaml`
