+++ services/berichtensessiecache/src/main/resources/application.properties
@@ -5,6 +5,15 @@ quarkus.application.version=0.1.0
 # HTTP
 quarkus.http.port=8080
 
+# Security headers op ÁLLE HTTP-responses (ook /openapi.json, /q/health,
+# /q/metrics, dev-UI) — niet alleen op JAX-RS-paden. De SecurityHeadersFilter
+# in fbs-common dekt JAX-RS responses; deze Quarkus-config dekt de rest.
+quarkus.http.header."X-Frame-Options".value=DENY
+quarkus.http.header."X-Content-Type-Options".value=nosniff
+quarkus.http.header."Strict-Transport-Security".value=max-age=31536000; includeSubDomains; preload
+quarkus.http.header."Content-Security-Policy".value=frame-ancestors 'none'
+quarkus.http.header."Referrer-Policy".value=no-referrer
+
 # OpenAPI / Swagger UI (ADR: spec MOET op /openapi.json beschikbaar zijn)
 quarkus.smallrye-openapi.path=/openapi.json
 quarkus.swagger-ui.always-include=true
